Change digits to be u32 instead of usize.
The signature of .pow() changed.

+++ src/lib.rs
@@ -20,7 +20,7 @@ fn dynamic_truncation(hs: &[u8]) -> u64 {
     p & 0x7fffffff
 }
 
-pub fn hotp_custom<D: Digest>(key: &[u8], counter: u64, digits: usize,
+pub fn hotp_custom<D: Digest>(key: &[u8], counter: u64, digits: u32,
                               hash: D) -> u64 {
     let mut hmac = Hmac::new(hash, key);
     u64_to_be_bytes(counter, 8, |bytes| {
@@ -32,32 +32,32 @@ pub fn hotp_custom<D: Digest>(key: &[u8], counter: u64, digits: usize,
     dynamic_truncation(hs) % 10.pow(digits)
 }
 
-pub fn hotp_raw(key: &[u8], counter: u64, digits: usize) -> u64 {
+pub fn hotp_raw(key: &[u8], counter: u64, digits: u32) -> u64 {
     let hash = Sha1::new();
     hotp_custom(key, counter, digits, hash)
 }
 
-pub fn hotp(key: &str, counter: u64, digits: usize) -> Result<u64, &str> {
+pub fn hotp(key: &str, counter: u64, digits: u32) -> Result<u64, &str> {
     match key.from_hex() {
         Ok(bytes) => Ok(hotp_raw(bytes.as_slice(), counter, digits)),
         Err(_) => Err("Unable to parse hex.")
     }
 }
 
-pub fn totp_custom<D: Digest>(key: &[u8], digits: usize, epoch: u64,
+pub fn totp_custom<D: Digest>(key: &[u8], digits: u32, epoch: u64,
                               time_step: u64, current_time: u64,
                               hash: D) -> u64 {
     let counter = (current_time - epoch) / time_step;
     hotp_custom(key, counter, digits, hash)
 }
 
-pub fn totp_raw(key: &[u8], digits: usize, epoch: u64, time_step: u64) -> u64 {
+pub fn totp_raw(key: &[u8], digits: u32, epoch: u64, time_step: u64) -> u64 {
     let hash = Sha1::new();
     let current_time = time::get_time();
     totp_custom(key, digits, epoch, time_step, current_time.sec as u64, hash)
 }
 
-pub fn totp(key: &str, digits: usize, epoch: u64,
+pub fn totp(key: &str, digits: u32, epoch: u64,
             time_step: u64) -> Result<u64, &str> {
     match key.from_hex() {
         Ok(bytes) => Ok(totp_raw(bytes.as_slice(), digits, epoch, time_step)),
